all tasks has been done

+++ database/migrations/2022_05_13_220658_create_affiliates_table.php
@@ -18,7 +18,12 @@ return new class extends Migration
             $table->foreignId('user_id');
             $table->foreignId('merchant_id');
             // TODO: Replace me with a brief explanation of why floats aren't the correct data type, and replace with the correct data type.
-            $table->float('commission_rate');
+
+            //Hammad: floats not recommended for storing currency or monetary values. The main reason is that floating-point numbers can lead to inaccuracies in calculations due to the way they are represented in binary format.
+
+            // Instead, it's better to use a precise data type that can accurately store monetary values without any loss of precision. In the context of currency values, it's common to use the DECIMAL data type.
+            
+            $table->decimal('commission_rate', 10, 2);
             $table->string('discount_code');
             $table->timestamps();
         });
