+++ database/migrations/2022_05_16_143445_create_orders_table.php
@@ -19,8 +19,13 @@ return new class extends Migration
             $table->foreignId('merchant_id')->constrained();
             $table->foreignId('affiliate_id')->nullable()->constrained();
             // TODO: Replace floats with the correct data types (very similar to affiliates table)
-            $table->float('subtotal');
-            $table->float('commission_owed')->default(0.00);
+
+           //Hammad: floats not recommended for storing currency or monetary values. The main reason is that floating-point numbers can lead to inaccuracies in calculations due to the way they are represented in binary format.
+
+            // Instead, it's better to use a precise data type that can accurately store monetary values without any loss of precision. In the context of currency values, it's common to use the DECIMAL data type.
+            
+            $table->decimal('subtotal', 10, 2);
+            $table->decimal('commission_owed', 10, 2)->default(0.00);
             $table->string('payout_status')->default(Order::STATUS_UNPAID);
             $table->string('discount_code')->nullable();
             $table->timestamps();
